% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{block_matrix}
\alias{block_matrix}
\title{Block matrix}
\usage{
block_matrix(x, step)
}
\arguments{
\item{x}{a matrix where the dimensions are both multiples of step}

\item{step}{the dimensions of the block matrix being selected over the diagonal of matrix x}
}
\description{
Block matrix
}
\examples{
# create a square matrix
m <- matrix(1:64, nrow=8)

# select a block matrix with step=2
block_matrix(m, step=2)

# select a block matrix with step=4
block_matrix(m, step=4)

# create a non-square matrix
nsm <- matrix(1:27, nrow=9, ncol=3 )

# check that both dimensions are multiples of step=3
dim(nsm) \%\% 3 == c(0, 0)

# select the diagonal step matrix
block_matrix(nsm, step=3)
}

