% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diag.R
\name{diag<-}
\alias{diag<-}
\title{Matrix Diagonals}
\usage{
diag(x) <- value
}
\arguments{
\item{x}{a matrix, vector or 1D array, or missing.}

\item{value}{either a single value or a vector of length equal to that of the current diagonal. Should be of a mode which can be coerced to that of x.}
}
\description{
A small extension of the `diag<-` function from the base package, allowing for replacement vectors smaller than the diagonal.
}
\details{
The `diag<-` function in the base package allows for two type of value, either of length 1, or of length equal to the length of the diagonal
This function extends that by allowing
}

