% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagonals.R
\docType{package}
\name{diagonals}
\alias{diagonals}
\alias{diagonals-package}
\title{diagonals}
\description{
diagonals
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
\seealso{
http://qua.st/diagonals
}

