% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matricise.R
\name{matricise}
\alias{matricise}
\title{Matricise}
\usage{
matricise(x, row_dim = NULL, col_dim = NULL)
}
\arguments{
\item{x}{a higher-order array (length(dim(x)) >= 3)}

\item{long}{should the output matrix be expanded along the rows, creating a long matrix, if FALSE then the output matrix is expanded along the columns, creating a wide matrix.}
}
\value{
a matrix (length(dim(x)) == 2 )
}
\description{
Matricise
}

