% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split_vector.R
\name{split_vector}
\alias{split_vector}
\title{Split Vector}
\usage{
split_vector(x, steps = NULL, size = NULL, replacement = 0)
}
\arguments{
\item{x}{a numeric or character vector}

\item{steps}{the number of steps}

\item{size}{the size of the step}

\item{replacement}{value to be inserted on the diagonal, by default this is
zero (0).}
}
\description{
Split Vector
}
\details{
Either steps or size is expected to be provided.
}

