% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{block_matrix}
\alias{block_matrix}
\title{Block matrix}
\usage{
block_matrix(x, steps = NULL, size = NULL)
}
\arguments{
\item{x}{a matrix where the dimensions are both multiples of size}

\item{steps}{the required number of steps (block matrices) across the diagonal}

\item{size}{the largest dimension of the matrix being selected over the diagonal of matrix}
}
\description{
Block matrix
}
\examples{
# create a square matrix
m <- matrix(1:64, nrow=8)

# select a block matrix with two steps
block_matrix(m, steps=2)

# select a block matrix with four steps
block_matrix(m, steps=4)

# create a non-square matrix
nsm <- matrix(1:27, nrow=9, ncol=3 )

# check that both dimensions are multiples of steps=3
dim(nsm) \%\% 3 == c(0, 0)

# select the diagonal matrix
block_matrix(nsm, steps=3)
}

