% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonals.R
\docType{package}
\name{diagonals}
\alias{diagonals}
\alias{diagonals-package}
\title{diagonals}
\description{
Several tools for handling block-matrix diagonals and similar
constructs are implemented. Block-diagonal matrices can be extracted or removed
using two small functions implemented here. In addition, non-square matrices
are supported. Block diagonal matrices occur when two dimensions of a data set
are combined along one edge of a matrix. For example, trade-flow data in the
decompr' and 'gvc' packages have each country-industry combination occur along
both edges of the matrix.
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
\seealso{
http://qua.st/diagonals
}

