% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fatdiag.R
\name{fatdiag}
\alias{fatdiag}
\alias{fatdiag<-}
\title{Fat Matrix Diagonals}
\usage{
fatdiag(x = 1, steps = NULL, size = NULL, nrow = NULL, ncol = NULL)

fatdiag(x, steps = NULL, size = NULL, on_diagonal = TRUE, byrow = TRUE)
  <- value
}
\arguments{
\item{x}{a square matrix where the dimensions are integer multiples of size or integer dividors of steps}

\item{steps}{the required number of steps (block matrices) across the diagonal}

\item{size}{the width or height of the matrix being dropped over the diagonal of matrix x}

\item{nrow}{the number of rows}

\item{ncol}{the number of columns}

\item{value}{replacement value}

\item{on_diagonal}{should the operation be apply to the elements on the fat diagonal.}

\item{byrow}{should replacement occur by row, alternatively this is done by column (only relevant for replacement vectors).}
}
\description{
Fat Matrix Diagonals
}
\details{
Either steps or size is expected to be provided.
}

