% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricise.R
\name{matricise}
\alias{matricise}
\title{Matricise}
\usage{
matricise(x, row_dim = c(NULL, 3, 4), col_dim = c(NULL, 3, 4))
}
\arguments{
\item{x}{a higher-order array (length(dim(x)) >= 3)}

\item{row_dim}{the input array dimension which should be added to the row dimension of the output matrix, the value has to be 3 or 4.}

\item{col_dim}{the input array dimension which should be added to the column dimension of the output matrix, the value has to be 3 or 4.}
}
\value{
a matrix (length(dim(x)) == 2 )
}
\description{
Matricise
}
