% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minus_block_matrix.R
\name{minus_block_matrix}
\alias{minus_block_matrix}
\title{Minus Block matrix}
\usage{
minus_block_matrix(x, steps = NULL, size = NULL)
}
\arguments{
\item{x}{a matrix where the dimensions are both multiples of size}

\item{steps}{the required number of steps (block matrices) across the diagonal}

\item{size}{the largest dimension of the matrix being dropped over the diagonal of matrix x}
}
\description{
Minus Block matrix
}
\examples{
# create a square matrix
m <- matrix(1:64, nrow=8)

# select a block matrix with size=2
minus_block_matrix(m, size=2)

# select a block matrix with size=4
minus_block_matrix(m, size=4)

# create a non-square matrix
nsm <- matrix(1:27, nrow=9, ncol=3 )

# check that both dimensions are multiples of size=3
dim(nsm) \%\% 3 == c(0, 0)

# select the diagonal size matrix
minus_block_matrix(nsm, size=3)
}

